lambda-container/requirements.txt
Keras
librosa
numba
numpy
mysql-connector-python
boto3
tensorflow
awslambdaric
